requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cloudinary==1.30.0
Django==4.1.5
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
